chore: hook for node auto select on spc search
